dtu_mlops_69/requirements.txt
torch==2.4
typer==0.15.1
fastapi==0.115.6
uvicorn==0.34.0
pytorch_lightning
numpy==1.26.4
kaggle==1.6.17
transformers==4.48.0
google-cloud-storage==2.19.0
pandas==2.2.3
scikit-learn==1.6.1
wandb==0.19.3
seaborn==0.13.2
matplotlib==3.10.0
pytest==8.3.4
requests==2.32.3
fastapi==0.115.6
python-multipart == 0.0.20
httpx==0.28.1
streamlit==1.41.1
pytest-cov==6.0.0
evidently==0.5.1
pep8==1.7.1